Move NewsAdapter to its own class
